Move active context methods to context.py
