+++ layer/context.py
@@ -7,9 +7,34 @@ from layer.tracker.ui_progress_tracker import RunProgressTracker
 from layer.training.base_train import BaseTrain
 
 
+# We store the active context temporarily so it can be used within
+# either a @layer decorated function or soon via `with layer.model() as context:`
+_ACTIVE_CONTEXT: Optional["Context"] = None
+
+
+def set_active_context(context: "Context") -> None:
+    global _ACTIVE_CONTEXT
+    _ACTIVE_CONTEXT = context
+
+
+def reset_active_context() -> None:
+    global _ACTIVE_CONTEXT
+    _ACTIVE_CONTEXT = None
+
+
+def get_active_context() -> Optional["Context"]:
+    """
+    Returns the active context object set from the active computation.
+    Used in local mode to identify which context to log resources to.
+
+    @return:  active context object
+    """
+    return _ACTIVE_CONTEXT
+
+
 class Context:
     """
-    Provides access to variables within the pipeline execution.
+    Provides access to variables within a given execution, for example during a model train.
 
     This class should not be initialized by end-users.
     """
