+++ layer/executables/entrypoint/model_trainer.py
@@ -12,6 +12,7 @@ from layerapi.api.ids_pb2 import ModelTrainId
 
 from layer import Context
 from layer.clients.layer import LayerClient
+from layer.context import reset_active_context, set_active_context
 from layer.contracts.assertions import Assertion
 from layer.contracts.asset import AssetType
 from layer.exceptions.exception_handler import exception_handler
@@ -21,11 +22,7 @@ from layer.exceptions.status_report import (
     ExecutionStatusReportFactory,
     PythonExecutionStatusReport,
 )
-from layer.global_context import (
-    current_project_full_name,
-    reset_active_context,
-    set_active_context,
-)
+from layer.global_context import current_project_full_name
 from layer.resource_manager import ResourceManager
 from layer.tracker.progress_tracker import RunProgressTracker
 from layer.training.train import Train
