+++ layer/main/asset.py
@@ -5,19 +5,19 @@ from layer.cache.utils import is_cached
 from layer.clients.layer import LayerClient
 from layer.config import ConfigManager
 from layer.config.config import Config
-from layer.context import Context
+from layer.context import (
+    Context,
+    get_active_context,
+    reset_active_context,
+    set_active_context,
+)
 from layer.contracts.assets import AssetPath, AssetType
 from layer.contracts.datasets import Dataset
 from layer.contracts.models import Model
 from layer.contracts.project_full_name import ProjectFullName
 from layer.contracts.tracker import ResourceTransferState
 from layer.exceptions.exceptions import ProjectInitializationException
-from layer.global_context import (
-    current_account_name,
-    get_active_context,
-    reset_active_context,
-    set_active_context,
-)
+from layer.global_context import current_account_name
 from layer.projects.utils import get_current_project_full_name
 from layer.tracker.utils import get_progress_tracker
 from layer.utils.async_utils import asyncio_run_in_thread
